+++ tycho-gpg-plugin/src/main/java/org/eclipse/tycho/gpg/SignRepositoryArtifactsMojo.java
@@ -10,16 +10,12 @@
 package org.eclipse.tycho.gpg;
 
 import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileWriter;
 import java.io.IOException;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
 import java.util.Arrays;
-import java.util.Enumeration;
+import java.util.Collection;
 import java.util.List;
-import java.util.jar.JarEntry;
-import java.util.jar.JarFile;
+import java.util.stream.Collectors;
+import java.util.stream.StreamSupport;
 
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugin.MojoFailureException;
@@ -30,24 +26,32 @@ import org.apache.maven.plugins.annotations.Parameter;
 import org.apache.maven.plugins.gpg.AbstractGpgMojoExtension;
 import org.apache.maven.plugins.gpg.ProxySignerWithPublicKeyAccess;
 import org.apache.maven.project.MavenProject;
+import org.bouncycastle.openpgp.PGPSignature;
 import org.codehaus.plexus.archiver.Archiver;
-import org.codehaus.plexus.archiver.ArchiverException;
 import org.codehaus.plexus.archiver.UnArchiver;
 import org.codehaus.plexus.archiver.xz.XZArchiver;
-import org.codehaus.plexus.archiver.xz.XZUnArchiver;
-import org.codehaus.plexus.archiver.zip.ZipArchiver;
 import org.codehaus.plexus.archiver.zip.ZipUnArchiver;
-import org.codehaus.plexus.util.xml.Xpp3Dom;
-import org.codehaus.plexus.util.xml.Xpp3DomBuilder;
-import org.codehaus.plexus.util.xml.Xpp3DomWriter;
-import org.codehaus.plexus.util.xml.pull.XmlPullParserException;
+import org.eclipse.equinox.internal.p2.artifact.processors.pgp.PGPSignatureVerifier;
+import org.eclipse.equinox.internal.p2.artifact.repository.simple.SimpleArtifactRepository;
+import org.eclipse.equinox.p2.core.IProvisioningAgent;
+import org.eclipse.equinox.p2.core.ProvisionException;
+import org.eclipse.equinox.p2.metadata.IArtifactKey;
+import org.eclipse.equinox.p2.repository.IRepository;
+import org.eclipse.equinox.p2.repository.artifact.ArtifactKeyQuery;
 import org.eclipse.equinox.p2.repository.artifact.IArtifactDescriptor;
+import org.eclipse.equinox.p2.repository.artifact.IArtifactRepositoryManager;
+import org.eclipse.equinox.p2.repository.artifact.IFileArtifactRepository;
+import org.eclipse.equinox.p2.repository.artifact.spi.ArtifactDescriptor;
+import org.eclipse.osgi.signedcontent.SignedContentFactory;
+import org.eclipse.sisu.equinox.EquinoxServiceFactory;
 
 /**
- * Modifies the p2 metadata (<code>artifacts.xml</code>) to add PGP signatures for each included
- * artifact. Signatures are added as <code>pgp.signatures</code> property on the artifact metadata,
- * in armored form; and public keys of the signers are optionally added as
- * <code>pgp.publicKeys</code> property on the repository metadata, in armored form.
+ * Modifies the p2 metadata ({@code artifacts.xml}) to add a PGP signature to each included
+ * artifact. A signature is added as a {@code pgp.signatures} property on the artifact metadata, in
+ * armored form; and the public key of the signer is {@link #addPublicKeyToRepo optionally added} as
+ * a {@code pgp.publicKeys} property on the repository metadata, in armored form, and/or
+ * {@link #addPublicKeyToRepo optionally added} as a {@code pgp.publicKeys} property on the artifact
+ * metadata, in armored form.
  * 
  * @see <a href=
  *      "https://help.eclipse.org/latest/index.jsp?topic=%2Forg.eclipse.platform.doc.isv%2Fguide%2Fp2_pgp.html">Using
@@ -55,6 +59,9 @@ import org.eclipse.equinox.p2.repository.artifact.IArtifactDescriptor;
  */
 @Mojo(name = "sign-p2-artifacts", requiresProject = true, defaultPhase = LifecyclePhase.PREPARE_PACKAGE)
 public class SignRepositoryArtifactsMojo extends AbstractGpgMojoExtension {
+    enum PGPKeyBehavior {
+        skip, replace, merge
+    }
 
     /**
      * The maven project.
@@ -62,38 +69,70 @@ public class SignRepositoryArtifactsMojo extends AbstractGpgMojoExtension {
     @Parameter(defaultValue = "${project}", readonly = true, required = true)
     protected MavenProject project;
 
+    /**
+     * The repository location.
+     */
     @Parameter(defaultValue = "${project.build.directory}/repository")
     private File repository;
 
     /**
-     * Configures to <code>true</code> to generate PGP signature only for artifacts that do
-     * <strong>not</strong> already contain signatures files from jarsigner.
+     * Configure to {@code true} to generate PGP signature only for artifacts that are
+     * <strong>not</strong> signed by a jarsigner.
      */
     @Parameter(defaultValue = "true")
     private boolean skipIfJarsigned;
 
     /**
-     * If set to {@code true} adds the public key belonging to the signature to the repository
+     * Configure to {@code true} to generate a PGP signature only for artifacts that do
+     * <strong>not</strong> contain a signature that's anchored in Java's trust store, i.e.,
+     * anchored in the JDK's cacerts. A JCA certificate, for example, is never anchored.
+     */
+    @Parameter(defaultValue = "true")
+    private boolean skipIfJarsignedAndAnchored;
+
+    /**
+     * Configure to {@code true} to generate a PGP signature for binary artifacts.
+     */
+    @Parameter(defaultValue = "true")
+    boolean skipBinaries;
+
+    /**
+     * Configure to {@code true} to add the public key of each signature to the repository's
      * metadata.
      */
     @Parameter(defaultValue = "true")
     private boolean addPublicKeyToRepo;
 
     /**
-     * If set to {@code true} adds the public key belonging to the signature to each signed
-     * artifact's metadata.
+     * Configure to {@code true} to add the public key of the signature to each signed artifact's
+     * metadata.
      */
     @Parameter(defaultValue = "true", alias = "addPublicKeysToArtifacts")
     private boolean addPublicKeyToArtifacts;
 
     /**
-     * Bundles that should be signed independently of other settings, e.g. {@link #skipIfJarsigned}.
+     * Configures how to generate PGP signatures for artifacts that already have one or more PGP
+     * signatures, {@code skip} to generate no new PGP signature, {@code replace} to replace the
+     * existing signature(s) with a new signature, and {@code merge} to add a new signature to any
+     * existing signature(s).
+     */
+    @Parameter(defaultValue = "skip")
+    private PGPKeyBehavior pgpKeyBehavior;
+
+    /**
+     * Configure the signer used for PGP signing. Currently supported are {@code gpg} for launching
+     * the native {@code gpg} executable, and {@code bc} for using Bouncy Castle libraries. The
+     * latter is much faster and it can sign in parallel, so is very much faster.
+     */
+    @Parameter(property = "tycho.pgp.signer", defaultValue = "gpg")
+    private String signer;
+
+    /**
+     * Configured to specify artifacts that should be signed independently of other settings, e.g.,
+     * {@link #skipIfJarsigned}, {@link #skipIfJarsignedAndAnchored}, and {@link #skipBinaries}.
      */
     @Parameter
     private List<String> forceSignature;
-
-    @Component(role = UnArchiver.class, hint = "xz")
-    private XZUnArchiver xzUnarchiver;
 
     @Component(role = UnArchiver.class, hint = "zip")
     private ZipUnArchiver zipUnArchiver;
@@ -101,142 +140,156 @@ public class SignRepositoryArtifactsMojo extends AbstractGpgMojoExtension {
     @Component(role = Archiver.class, hint = "xz")
     private XZArchiver xzArchiver;
 
-    @Component(role = Archiver.class, hint = "zip")
-    private ZipArchiver zipArchiver;
+    @Component(role = IProvisioningAgent.class)
+    private IProvisioningAgent agent;
+
+    @Component(role = EquinoxServiceFactory.class, hint = "tycho-core")
+    private EquinoxServiceFactory serviceFactory;
+
+    @Override
+    protected String getSigner() {
+        return signer;
+    }
+
+    @Override
+    protected File getPGPInfo() {
+        var pgpInfo = System.getProperty("org.eclipse.tycho.test.pgp.info");
+        return pgpInfo == null ? null : new File(pgpInfo);
+    }
 
     @Override
     public void execute() throws MojoExecutionException, MojoFailureException {
-        var artifactsXml = new File(repository, "artifacts.xml");
-        var artifactsXmlXz = new File(repository, "artifacts.xml.xz");
-        var artifactsJar = new File(repository, "artifacts.jar");
-        if (!artifactsXml.exists()) {
-            if (artifactsXmlXz.exists()) {
-                xzUnarchiver.setSourceFile(artifactsXmlXz);
-                xzUnarchiver.setDestFile(artifactsXml);
-                xzUnarchiver.extract();
-                artifactsXmlXz.delete();
-            }
-            if (artifactsJar.exists()) {
-                zipUnArchiver.setSourceFile(artifactsJar);
-                zipUnArchiver.setDestDirectory(repository);
-                zipUnArchiver.extract();
-                artifactsJar.delete();
-            }
-        }
-        Xpp3Dom dom = null;
-        try (var stream = new FileInputStream(artifactsXml)) {
-            dom = Xpp3DomBuilder.build(stream, StandardCharsets.UTF_8.displayName());
-        } catch (IOException | XmlPullParserException e) {
-            throw new MojoExecutionException(e.getMessage(), e);
-        }
-        ProxySignerWithPublicKeyAccess signer = newSigner(project);
-        String armoredPublicKey = signer.getPublicKeys();
-        for (var artifact : dom.getChild("artifacts").getChildren("artifact")) {
-            Xpp3Dom properties = artifact.getChild("properties");
-            if (Arrays.stream(properties.getChildren())
-                    .anyMatch(property -> IArtifactDescriptor.FORMAT.equals(property.getAttribute("name"))
-                            && IArtifactDescriptor.FORMAT_PACKED.equals(property.getAttribute("value")))) {
-                // skip packed artifacts
-                continue;
-            }
-            /*
-             * Different types of artifact have different locations in the repo. So we need to check
-             * the classifier to correctly get the location of the jar file.
-             * 
-             * Could potentially do this dynamically based on the mappings attribute with the repo.
-             */
-            String subDir = null;
-            switch (artifact.getAttribute("classifier")) {
-            case "osgi.bundle":
-                subDir = "plugins";
-                break;
-            case "org.eclipse.update.feature": //Not yet signing features
-            case "binary": //Not yet signing binaries
-            default:
-                continue; // Skip signing
-            }
-            var file = new File(repository, subDir + File.separator + artifact.getAttribute("id") + '_'
-                    + artifact.getAttribute("version") + ".jar");
-            if (!file.canRead()) {
-                continue;
-            }
-            if (skipSign(file)) {
-                continue;
-            }
-            var signatureFile = signer.generateSignatureForArtifact(file);
-            try {
-                String signature = Files.readString(signatureFile.toPath());
-                var signatureProperty = new Xpp3Dom("property");
-                signatureProperty.setAttribute("name", "pgp.signatures");
-                signatureProperty.setAttribute("value", signature);
-                properties.addChild(signatureProperty);
-                properties.setAttribute("size",
-                        Integer.toString(Integer.parseInt(properties.getAttribute("size")) + 1));
-                if (addPublicKeyToArtifacts) {
-                    var publicKeyProperty = new Xpp3Dom("property");
-                    publicKeyProperty.setAttribute("name", "pgp.publicKeys");
-                    publicKeyProperty.setAttribute("value", armoredPublicKey);
-                    properties.addChild(publicKeyProperty);
-                    properties.setAttribute("size",
-                            Integer.toString(Integer.parseInt(properties.getAttribute("size")) + 1));
-                }
-            } catch (IOException e) {
-                throw new MojoExecutionException(e.getMessage(), e);
-            }
-        }
-        if (addPublicKeyToRepo) {
-            Xpp3Dom repositoryProperties = dom.getChild("properties");
-            repositoryProperties.setAttribute("size",
-                    Integer.toString(Integer.parseInt(repositoryProperties.getAttribute("size")) + 1));
-            var signersProperty = new Xpp3Dom("property");
-            signersProperty.setAttribute("name", "pgp.publicKeys");
-            signersProperty.setAttribute("value", armoredPublicKey);
-            repositoryProperties.addChild(signersProperty);
-        }
-        try (var writer = new FileWriter(artifactsXml)) {
-            Xpp3DomWriter.write(writer, dom);
-        } catch (IOException e) {
-            throw new MojoExecutionException(e.getMessage(), e);
-        }
-        xzArchiver.setDestFile(artifactsXmlXz);
-        xzArchiver.addFile(artifactsXml, artifactsXml.getName());
+        var signedContentFactory = serviceFactory.getService(SignedContentFactory.class);
+
+        var signer = newSigner(project);
+        var keys = KeyStore.create();
+
+        var artifactRepositoryManager = agent.getService(IArtifactRepositoryManager.class);
         try {
+            var artifactRepository = (IFileArtifactRepository) artifactRepositoryManager
+                    .loadRepository(repository.toURI(), null);
+
+            var compressed = "true".equals(artifactRepository.getProperty(IRepository.PROP_COMPRESSED));
+
+            if (addPublicKeyToRepo && this.pgpKeyBehavior != PGPKeyBehavior.replace) {
+                var repositoryKeys = artifactRepository.getProperty(PGPSignatureVerifier.PGP_SIGNER_KEYS_PROPERTY_NAME);
+                if (repositoryKeys != null) {
+                    keys.add(repositoryKeys);
+                }
+            }
+
+            var artifactKeys = artifactRepository.query(ArtifactKeyQuery.ALL_KEYS, null);
+            var descriptors = StreamSupport.stream(artifactKeys.spliterator(), false)
+                    .map(artifactRepository::getArtifactDescriptors).map(Arrays::asList).flatMap(Collection::stream)
+                    .collect(Collectors.toList());
+            descriptors.parallelStream().forEach(
+                    it -> handle(it, artifactRepository.getArtifactFile(it), signedContentFactory, signer, keys));
+
+            if (addPublicKeyToRepo && !keys.isEmpty()) {
+                artifactRepository.setProperty(PGPSignatureVerifier.PGP_SIGNER_KEYS_PROPERTY_NAME,
+                        keys.toArmoredString());
+            }
+
+            ((SimpleArtifactRepository) artifactRepository).save();
+
+            var artifactsXml = new File(repository, "artifacts.xml");
+            var artifactsXmlXz = new File(repository, "artifacts.xml.xz");
+            var artifactsJar = new File(repository, "artifacts.jar");
+            if (!artifactsXml.exists()) {
+                if (artifactsJar.exists()) {
+                    zipUnArchiver.setSourceFile(artifactsJar);
+                    zipUnArchiver.setDestDirectory(repository);
+                    zipUnArchiver.extract();
+                }
+            }
+
+            xzArchiver.setDestFile(artifactsXmlXz);
+            xzArchiver.addFile(artifactsXml, artifactsXml.getName());
             xzArchiver.createArchive();
-        } catch (ArchiverException | IOException e) {
-            throw new MojoExecutionException(e.getMessage(), e);
-        }
-        zipArchiver.setDestFile(artifactsJar);
-        zipArchiver.addFile(artifactsXml, artifactsXml.getName());
-        try {
-            zipArchiver.createArchive();
-        } catch (ArchiverException | IOException e) {
-            throw new MojoExecutionException(e.getMessage(), e);
-        }
-    }
-
-    private boolean skipSign(File file) throws MojoFailureException {
-        if (forceSignature != null) {
-            String bundleName = file.getName().substring(0, file.getName().lastIndexOf('_'));
-            if (forceSignature.contains(bundleName)) {
-                return false;
+            if (compressed) {
+                artifactsXml.delete();
             }
+
+        } catch (RuntimeException e) {
+            var cause = e.getCause();
+            if (cause instanceof MojoExecutionException) {
+                throw (MojoExecutionException) cause;
+            }
+            if (cause instanceof MojoFailureException) {
+                throw (MojoFailureException) cause;
+            }
+            throw e;
+        } catch (ProvisionException | IOException e) {
+            throw new MojoExecutionException(e.getMessage(), e);
         }
-        return skipIfJarsigned && isJarSigned(file);
     }
 
-    private boolean isJarSigned(File file) throws MojoFailureException {
-        try (JarFile jar = new JarFile(file)) {
-            Enumeration<JarEntry> entries = jar.entries();
-            while (entries.hasMoreElements()) {
-                String name = entries.nextElement().getName();
-                if (name.startsWith("META-INF/") && name.endsWith(".SF")) {
-                    return true;
+    private void handle(IArtifactDescriptor artifactDescriptor, File artifact,
+            SignedContentFactory signedContentFactory, ProxySignerWithPublicKeyAccess signer, KeyStore allKeys) {
+        if (artifact != null) {
+            var existingKeys = artifactDescriptor.getProperty(PGPSignatureVerifier.PGP_SIGNER_KEYS_PROPERTY_NAME);
+            var existingSignatures = artifactDescriptor.getProperty(PGPSignatureVerifier.PGP_SIGNATURES_PROPERTY_NAME);
+
+            if (existingSignatures != null && pgpKeyBehavior == PGPKeyBehavior.skip) {
+                return;
+            }
+
+            IArtifactKey artifactKey = artifactDescriptor.getArtifactKey();
+
+            if (forceSignature == null || !forceSignature.contains(artifactKey.getId())) {
+                var classifier = artifactKey.getClassifier();
+                var isBinary = "binary".equals(classifier);
+                if (skipBinaries && isBinary) {
+                    return;
+                }
+
+                if (!isBinary) {
+                    var service = serviceFactory.getService(SignedContentFactory.class);
+                    try {
+                        var signedContent = service.getSignedContent(artifact);
+                        if (signedContent.isSigned()) {
+                            if (skipIfJarsigned) {
+                                return;
+                            }
+                            if (skipIfJarsignedAndAnchored) {
+                                for (var signerInfo : signedContent.getSignerInfos()) {
+                                    if (signerInfo.getTrustAnchor() != null) {
+                                        return;
+                                    }
+                                }
+                            }
+                        }
+                    } catch (Exception e) {
+                        //$FALL-THROUGH$ Treat as unsigned.
+                    }
                 }
             }
-        } catch (IOException e) {
-            throw new MojoFailureException(e);
-        }
-        return false;
-    }
 
+            if (pgpKeyBehavior == PGPKeyBehavior.replace) {
+                existingSignatures = null;
+                existingKeys = null;
+            }
+
+            try {
+                var signatures = signer.generateSignature(artifact);
+                var signerKeys = signatures.all().stream().map(PGPSignature::getKeyID)
+                        .flatMap(id -> signer.getPublicKeys().getKeys(id).stream()).collect(Collectors.toList());
+                var keyStore = KeyStore.create(existingKeys);
+                keyStore.add(signerKeys);
+                allKeys.add(keyStore);
+
+                signatures.add(existingSignatures);
+
+                ((ArtifactDescriptor) artifactDescriptor).setProperty(PGPSignatureVerifier.PGP_SIGNATURES_PROPERTY_NAME,
+                        signatures.toArmoredString());
+
+                if (addPublicKeyToArtifacts) {
+                    ((ArtifactDescriptor) artifactDescriptor).setProperty(
+                            PGPSignatureVerifier.PGP_SIGNER_KEYS_PROPERTY_NAME, keyStore.toArmoredString());
+                }
+            } catch (MojoExecutionException | IOException e) {
+                throw new RuntimeException(e.getMessage(), e);
+            }
+        }
+    }
 }